Replaced SDL_Renderer with SpriteRenderer

+++ source/Timer.h
@@ -11,7 +11,7 @@ struct Timer {
   uint32_t msElapsed;
   uint32_t accumulator;
   float dt;             //multiply times movement to get pixels per second
-  uint8_t MSPerUpdate;  //TODO: decide on whether this can be changed live
+  float MSPerUpdate;  //TODO: decide on whether this can be changed live
 };
 
 void updateTimer(Timer* timer)
